FIX: improvements after code review
