+++ app/assets/stylesheets/common/base/_index.scss
@@ -47,6 +47,7 @@
 @import "share_link";
 @import "shared-drafts";
 @import "sidebar";
+@import "sidebar-custom-section";
 @import "sidebar-footer";
 @import "sidebar-section";
 @import "sidebar-more-section-links";
