+++ app/assets/stylesheets/common/base/sidebar.scss
@@ -128,40 +128,6 @@
   }
 }
 
-.sidebar-custom-sections {
-  .sidebar-section-wrapper {
-    padding-bottom: 0;
-  }
-  .d-icon-globe {
-    position: absolute;
-    left: 0.5em;
-    height: 0.75em;
-    width: 0.75em;
-    margin-top: 0.15em;
-    align-items: center;
-  }
-
-  .sidebar-section-link-prefix.icon {
-    cursor: move;
-  }
-
-  .sidebar-section-wrapper.disabled {
-    .sidebar-section-link-wrapper {
-      .sidebar-section-link-prefix.icon,
-      .sidebar-section-link {
-        background: none;
-        color: var(--primary-low);
-      }
-      .sidebar-section-link.drag {
-        font-weight: bold;
-        color: var(--primary-high);
-        .sidebar-section-link-prefix.icon {
-          color: var(--primary-high);
-        }
-      }
-    }
-  }
-}
 .sidebar-section-form-modal {
   .modal-inner-container {
     width: var(--modal-max-width);
