+++ app/serializers/sidebar_section_serializer.rb
@@ -4,7 +4,7 @@ class SidebarSectionSerializer < ApplicationSerializer
   attributes :id, :title, :links, :slug, :public
 
   def links
-    object.sidebar_section_links.sort_by(&:position).map(&:linkable)
+    object.sidebar_section_links.order(position: :asc).map(&:linkable)
   end
 
   def slug
