+++ db/migrate/20230307051200_add_position_to_sidebar_section_links.rb
@@ -3,5 +3,8 @@
 class AddPositionToSidebarSectionLinks < ActiveRecord::Migration[7.0]
   def change
     add_column :sidebar_section_links, :position, :integer, default: 0, null: false
+    add_index :sidebar_section_links,
+              %i[user_id sidebar_section_id position],
+              name: "links_user_id_section_id_position"
   end
 end
